Add code to cache the json into a separate table in the database and accompanying code to use said data so that ffprobe doesn't need to be executed every time the videoinfo table is regenerated.
Add argument to delete all entries from the videojson table and exit.

Move createstatement variables to near the top of the file (no reason not to).

Use proper variable for the individual entry deletion and deleting a file from the videoinfo table will also delete that file from the videojson table (It will just cache the json again if genvideoinfo is used on it anyways)