Move ABS class out of videoutil module (clogging up the file unnecessarily) (commit 1 of 2 for this).
Make GenVideoInfo use the abstraction functions.

Make VideoInfo class independent of the VideoUtil class as it was causing a dependency loop and the variables that it was using are for connecting to the videoinfo database anyway.

Remove FindVideoInfo's sub-classing from VideoInfo because its not using any of the functions or variables from that class.

Remove database related parameters from VideoUtil.

Add ffprobe existence query to VideoInfo because its no longer getting that info from VideoUtil.