Add options to disable videoinfo generation and change the name/location of the output database.
Moved all the GenVideoInfo related operations to a function.

Output directory will now be created if it doesn't exist, it will error out on <python 3.5 if (for some reason) the directory exists and pathlib didn't do its job.

If the output directory is not the current directory, the code to generate video info will now create a database in the output directory (or at the location specified in --output-database) and put the information there.