Major rework of VideoInfo.find.
Make a few statements more verbose and make a few statements only print if their function's verbose mode is on.

Make VideoInfo::Generate.existing more robust in the case that the database is totally empty, fix the query in VideoInfo::Generate.json to return the json data instead of the file name (it was originally supposed to be a check, but turned into the variable where the data was supposed to go later on. Add a few respond_to? checks for some streams[1] statements in VideoInfo::Generate.hash, will do more if necessary.