Fix variable for --reset-all. Also:
Change file discovery to use Find instead of Pathname.descend (wording on what that does makes me think that there could potentially be weird edge cases that would annoy me).

Use Array.delete_if instead of code in VideoInfo::Generate.filelist (as of this commit, this script is the only one using that code anyway).

Switch to Util::HashFile.genhash instead of the redundant version in VideoInfo::Generate (it was just calling the Util function and creating a hash within a hash).

Move text about extracting metadata to VideoInfo::Generate.json and make it only come up on a cache miss, a separate message has been added for a cache hit.