Remove some unnecessary code (was from a previous hack that turned out to not be the right hack).
Add raise to the rescue statement in the json generation function, but only if verbose is specified because verbose mode is where all helpful info is, will just print a warning message if verbose mode is not enabled.

Add argument to json function to take a repo symbol so that it ensures that the table name is accurate no matter what database is being used (theoretically).

Use the filepath.realpath instead of filename variable because convertmkv was somehow giving it a filename that it couldn't find in the filehash hash when it was outputting files to a different directory using a secondary database.