Clean up andy/git.py a bit.
Move url from class initialization to the clone function since that's the only function that uses it.

Change the type check to use isinstance instead of an x in y check. Also, raise TypeError instead of ValueError since the former is more appropriate.

Put the lock path object in a variable instead of having it written out multiple times for no reason.

Don't use a deque for gccmd in the gc function because its not necessary.