Add URL verification
Make date affixing (and the no_date option) only apply to the default directory. If you're specifying a custom directory, adding the date yourself shouldn't be a problem.

Remove commented out code

Make sure to exit gracefully if there are no entries in the Files constant or if said constant is nil.