Reorganize methods so that the module methods are all on top.
Properly handle a pre-parsed hash in GenerateVideoInfo.genhash

Change the frame rate calculation to use a case statement and have the calculator invocation be a lambda function since it was the same call just different sources.

Change the respond_to? queries to add the question mark for :key? in bitrate_1_raw

Remove unnecessary code block.

Change the filelist whitelist to simply concatenate the two lists together, that's really all that needs to be done.