Change '.config' + 'convertmkv.json' to just '.config/convertmkv.json' for the default argument for --convertmkv-config. No need to do the extra append operation.
Add code to download using ffmpeg, I'm hoping that it will help with a certain corner case.

Change the archive code to be a local variable instead of a constant since its not needed outside the invocation of youtube-dl

Rename whitelist to blacklist in the muxfiles.delete_if statement since that is what it is.