Change privuser and value to case statements in Util.privileged?
Change than to a case statement in Util::DateDiff.getdiff to a case statement and fix the logic.

Revert the optional program variable logic in Util::Program.runprogram since it already had code to work with blocks, but for utilizing the output instead of determining the command line. Add code to convert non-array program objects to arrays if they can be, but the conversion will not be run if program is already an array. Add Subprocess::NonZeroExit to the rescue statement in Util::Program.runprogram.