Add option to force run garbage collection
Rename the ttrssgcfile method to touchgcfile and use it for touching the GC file if garbage collection is run.

Fix unconditional touching of the GC file

Change raise statement in the Args.replace case statement to just use a raise statement with a "reason" instead of printing a Mood.sad statement.