Convert Util::SortEntries.sort to Util.sort
Use a keep_if statement for culling the file list of non-video files instead of the hack i was using before.

For the test mode, make a Hash that has the filename as a key and a sub-hash that has the mime type and whether the file's mime type is in the whitelist as the data.