Add a method called cleanup_array which will flatten, compact, and dedup (if the unique keyword is true) the array given to it.
Change hashfile method to use a lambda to do the actual hash calculation since the code for both sides of the if/else were doing basically the same thing, only difference being whether it was done to one file or many. Also, this means that it won't reinstantiate OpenSSL::Digest every time.