Use only needed util modules
Switch to videoinfo module

Use cleanup! method instead of the old boilerplate code i used to do for each array.

Change Command class to cmdline method, there was no reason for it to be a class, i think i did it that way as a practice for doing accessors.

Go back to << instead of += for all the statements adding to the main array in cmdline because I can't necessarily guarantee that all the statements going into the main array will be arrays themselves, the cleanup! method will take care of making sure that it comes out as a single array.