Switch to modularized util/core/videoinfo
Added toggle to use webm instead of mkv for videos with separate files per stream

Renamed options.no_stats to options.stats, it still is a negative action, but this method works better with some of the code.

Uncommented the mux code, but its now a positive action instead of a negative one (translation: you have to explicitly enable it instead of it being on by default)

Remove the exit actions and added conditionals so that the app will actually get to the muxing code if the array for the statistics calculation code is empty.