Many changes to ytdl, see extended commit message.
Change Args.no_date to Args.datesubdir and flip the if/unless logic

Convert Args.directory calculation to a lambda

Add code to blacklist files from the playlist which also means that there is now code to extract the filenames for the downloaded videos (currently only used to make the playlist blacklist, but could be used elsewhere in the future)

Convert the stats and muxing code to two separate methods each (one for calculating the file list and one to actually execute the file list.

More apostrophe/single quote escaping.