A few different changes, see extended commit note:
Rename the no playlist file from no-playlist.txt to .noplaylist

Make the blacklist code into a method and run it at the end of genplaylist so that if the playlist is deleted (but --reset-playlist is not enabled) it will not put older files into the new playlist. Uses a constant to indicate if blacklist has been run already so as not to run it twice. Also, remove the reference to being tied to --no-playlist.