Big changes in absconvert{/metadata,/cmdline}
Add power_p as an optional dependancy

Change from using a new variable to pretify argument queries to using
the bracket method in the ABSConvert::Options class

Remove the bitrate! and framerate! methods and have all that code
executed in initialize

Don't run ABSConvert.backup if debug mode is active.