Add code to work around pwd not being present.
It shouldn't be a problem on unix systems where this module will be
used, but you never know. is_privileged() will return None if pwd is not
present and we're not checking permissions on a directory. This is so
that I can throw an exception if any check returns None.