Move which method of FindApp to base Util module.
Leaving FindApp.which behind with a warning that its deprecated. It will
still run the code as it used to because it calls Util.findapp. When I
feel that all users have moved away, I will remove the class, obviously
any users that I missed will show up at that point.